+++ src/main/java/com/atlauncher/gui/tabs/tools/ToolsTab.java
@@ -36,7 +36,7 @@ public class ToolsTab extends JPanel implements Tab {
 
         JPanel mainPanel = new JPanel();
 
-        mainPanel.setLayout(new GridLayout(3, 2, 10, 10));
+        mainPanel.setLayout(new GridLayout(4, 2, 10, 10));
 
         final IToolsViewModel viewModel = new ToolsViewModel();
         mainPanel.add(new NetworkCheckerToolPanel(viewModel));
@@ -45,6 +45,7 @@ public class ToolsTab extends JPanel implements Tab {
         mainPanel.add(new RuntimeDownloaderToolPanel(viewModel));
         mainPanel.add(new DownloadClearerToolPanel(viewModel));
         mainPanel.add(new SkinUpdaterToolPanel(viewModel));
+        mainPanel.add(new LibrariesDeleterToolPanel(viewModel));
 
         add(mainPanel, BorderLayout.CENTER);
     }