+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 This changelog only contains the changes that are unreleased. For changes for individual releases, please visit the
 [releases](https://github.com/ATLauncher/ATLauncher/releases) page on GitHub.
 
-## 3.4.21.1
+## 3.4.22.0
 
 ### New Features
+- Add in new tool for deleting all libraries
 
 ### Fixes